+++ NGitLab.Mock/Clients/ClusterClient.cs
@@ -6,7 +6,7 @@ namespace NGitLab.Mock.Clients;
 
 internal sealed class ClusterClient : ClientBase, IClusterClient
 {
-    private readonly int _projectId;
+    private readonly long _projectId;
 
     public ClusterClient(ClientContext context, ProjectId projectId)
         : base(context)
