+++ NGitLab.Mock/Clients/GroupSearchClient.cs
@@ -6,7 +6,7 @@ namespace NGitLab.Mock.Clients;
 internal sealed class GroupSearchClient : ClientBase, ISearchClient
 {
     private readonly ClientContext _context;
-    private readonly int _groupId;
+    private readonly long _groupId;
 
     public GroupSearchClient(ClientContext context, GroupId groupId)
         : base(context)
