+++ NGitLab.Mock/GroupHookCollection.cs
@@ -23,7 +23,7 @@ public class GroupHookCollection : Collection<GroupHook>
         base.Add(item);
     }
 
-    private int GetNewId()
+    private long GetNewId()
     {
         return this.Select(hook => hook.Id).DefaultIfEmpty().Max() + 1;
     }
