+++ NGitLab/IEnvironmentClient.cs
@@ -33,7 +33,7 @@ public interface IEnvironmentClient
     /// - Renaming an environment with the API removed in GitLab 16.0.
     /// </remarks>
     [EditorBrowsable(EditorBrowsableState.Never)]
-    EnvironmentInfo Edit(int environmentId, string name, string externalUrl);
+    EnvironmentInfo Edit(long environmentId, string name, string externalUrl);
 
     /// <summary>
     /// Updates an existing environment's external_url.
@@ -41,20 +41,20 @@ public interface IEnvironmentClient
     /// <param name="environmentId">The ID of the environment</param>
     /// <param name="externalUrl">The new external url</param>
     /// <returns>The updated environment</returns>
-    EnvironmentInfo Edit(int environmentId, string externalUrl);
+    EnvironmentInfo Edit(long environmentId, string externalUrl);
 
     /// <summary>
     /// Delete an environment.
     /// </summary>
     /// <param name="environmentId">The ID of the environment</param>
-    void Delete(int environmentId);
+    void Delete(long environmentId);
 
     /// <summary>
     /// Stop an environment.
     /// </summary>
     /// <param name="environmentId">The ID of the environment</param>
     /// <returns>The stopped environment</returns>
-    EnvironmentInfo Stop(int environmentId);
+    EnvironmentInfo Stop(long environmentId);
 
     /// <summary>
     /// List environments of the project
@@ -68,12 +68,12 @@ public interface IEnvironmentClient
     /// </summary>
     /// <param name="environmentId">The ID of the environment</param>
     /// <returns>The environment with the corresponding ID</returns>
-    EnvironmentInfo GetById(int environmentId);
+    EnvironmentInfo GetById(long environmentId);
 
     /// <summary>
     /// Get a specific environment
     /// </summary>
     /// <param name="environmentId">The ID of the environment</param>
     /// <returns>The environment with the corresponding ID</returns>
-    Task<EnvironmentInfo> GetByIdAsync(int environmentId, CancellationToken cancellationToken = default);
+    Task<EnvironmentInfo> GetByIdAsync(long environmentId, CancellationToken cancellationToken = default);
 }
