+++ NGitLab/IGroupBadgeClient.cs
@@ -7,11 +7,11 @@ public interface IGroupBadgeClient
 {
     IEnumerable<Badge> All { get; }
 
-    Badge this[int id] { get; }
+    Badge this[long id] { get; }
 
     Badge Create(BadgeCreate badge);
 
-    Badge Update(int id, BadgeUpdate badge);
+    Badge Update(long id, BadgeUpdate badge);
 
-    void Delete(int id);
+    void Delete(long id);
 }
