+++ NGitLab/ISystemHookClient.cs
@@ -7,9 +7,9 @@ public interface ISystemHookClient
 {
     IEnumerable<SystemHook> All { get; }
 
-    SystemHook this[int hookId] { get; }
+    SystemHook this[long hookId] { get; }
 
     SystemHook Create(SystemHookUpsert hook);
 
-    void Delete(int hookId);
+    void Delete(long hookId);
 }
